feat: Swipe to delete
* feat: Swipe to delete
* add swiping delete functionality

Issue #86
(ref: https://github.com/COSC481W-2022Fall/capstone-project-vibe/issues/86)
